Move keycodes.js to src/helpers
